Fix: minor

+++ setup.sh
@@ -4,5 +4,10 @@ DOT_FILES=( .zshrc .zshenv .tmux.conf .vimrc .vim )
 
 for file in ${DOT_FILES[@]}
 do
-  ln -s $HOME/.dotfiles/$file $HOME/$file
+  if [ ! -e "$HOME/$file" ]; then
+    ln -s $HOME/.dotfiles/$file $HOME/$file
+  fi
 done
+
+ln -s ~/.vim ~/.config/nvim
+ln -s ~/.vimrc ~/.config/nvim/init.vim
